benchmarks/harness/pyproject.toml
[project]
name = "formualizer-comparative-bench-harness"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["pyyaml>=6.0"]

[tool.uv]
package = false
